Reverted some code back for 1.13.
I reverted the 1.14 only code back for 1.13 as I am waiting for a more stable release of 1.14 but I kept the code just commented it out.
