+++ src/main/java/me/badbones69/crazycrates/Main.java
@@ -46,11 +46,11 @@ public class Main extends JavaPlugin implements Listener {
 	
 	@Override
 	public void onEnable() {
-		if(Version.getCurrentVersion().isOlder(Version.v1_14_R1)) {// Disables plugin on 1.12.2 and below servers.
+		if(Version.getCurrentVersion().isOlder(Version.v1_13_R1)) {// Disables plugin on 1.12.2 and below servers.
 			isEnabled = false;
 			System.out.println("============= Crazy Crates =============");
 			System.out.println(" ");
-			System.out.println("Plugin Disabled: This server is running on 1.13.2 or below and Crazy Crates does not support those versions. "
+			System.out.println("Plugin Disabled: This server is running on 1.12.2 or below and Crazy Crates does not support those versions. "
 			+ "Please check the spigot page for more information about lower Minecraft versions.");
 			System.out.println(" ");
 			System.out.println("Plugin Page: https://www.spigotmc.org/resources/17599/");
