+++ src/main/resources/plugin.yml
@@ -2,7 +2,7 @@ name: CrazyCrates
 main: me.badbones69.crazycrates.Main
 author: BadBones69
 version: ${project.version}
-api-version: 1.14
+api-version: 1.13
 softdepend: [MVdWPlaceholderAPI, PlaceholderAPI]
 description: Want some good quality crates without paying money for them? Then check this crate plugin out.
 commands:
